Fix env-backed path resolution and config hashing
