+++ src/datapipeline/services/bootstrap/config.py
@@ -1,25 +1,15 @@
-import re
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Mapping
 
 from datapipeline.config.project import ProjectConfig
+from datapipeline.services.config_refs import (
+    interpolate_config_vars,
+    project_vars_from_data,
+    resolve_config_refs,
+    serialize_project_value,
+)
 from datapipeline.services.path_policy import resolve_project_path
 from datapipeline.utils.load import load_yaml
-from datapipeline.utils.placeholders import MissingInterpolation, is_missing
-from .refs import resolve_config_refs
-
-
-def _serialize_global_value(value: Any) -> Any:
-    """Normalize project global values for interpolation."""
-    if isinstance(value, datetime):
-        try:
-            return value.astimezone(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-        except ValueError:
-            return value.isoformat()
-    if value is None:
-        return None
-    return str(value)
 
 
 def _project(project_yaml: Path) -> ProjectConfig:
@@ -28,7 +18,7 @@ def _project(project_yaml: Path) -> ProjectConfig:
     vars_ = _project_vars(data)
     paths = data.get("paths")
     if isinstance(paths, dict) and vars_:
-        data["paths"] = _interpolate(paths, vars_)
+        data["paths"] = interpolate_config_vars(paths, vars_)
     return ProjectConfig.model_validate(data)
 
 
@@ -38,21 +28,7 @@ def _paths(project_yaml: Path) -> Mapping[str, str]:
 
 
 def _project_vars(data: dict) -> dict[str, Any]:
-    vars_: dict[str, Any] = {}
-    name = data.get("name")
-    if name:
-        vars_["project"] = str(name)
-        vars_["project_name"] = str(name)
-
-    version = data.get("version")
-    if version is not None:
-        vars_["version"] = str(version)
-        vars_["project_version"] = str(version)
-
-    globals_ = data.get("globals") or {}
-    for k, v in globals_.items():
-        vars_[str(k)] = _serialize_global_value(v)
-    return vars_
+    return project_vars_from_data(data)
 
 
 def artifacts_root(project_yaml: Path) -> Path:
@@ -121,42 +97,12 @@ def _globals(project_yaml: Path) -> dict[str, Any]:
     g = proj.globals.model_dump()
     out: dict[str, Any] = {}
     for k, v in g.items():
-        out[str(k)] = _serialize_global_value(v)
+        out[str(k)] = serialize_project_value(v)
     return out
 
 
-_VAR_RE = re.compile(r"\$\{([^}]+)\}")
-
-
 def _interpolate(obj, vars_: dict[str, Any]):
-    """Recursively substitute ${var} in strings using vars_ map.
-
-    Minimal behavior: if a key is missing, leave placeholder as-is.
-    """
-    if isinstance(obj, dict):
-        return {k: _interpolate(v, vars_) for k, v in obj.items()}
-    if isinstance(obj, list):
-        return [_interpolate(v, vars_) for v in obj]
-    if isinstance(obj, str):
-        match = _VAR_RE.fullmatch(obj)
-        if match:
-            key = match.group(1)
-            if key in vars_:
-                value = vars_[key]
-                if value is None or is_missing(value):
-                    return MissingInterpolation(key)
-                return str(value)
-            return obj
-
-        def repl(m):
-            key = m.group(1)
-            value = vars_.get(key, m.group(0))
-            if value is None or is_missing(value):
-                return m.group(0)
-            return str(value)
-
-        return _VAR_RE.sub(repl, obj)
-    return obj
+    return interpolate_config_vars(obj, vars_)
 
 
 __all__ = [
