+++ tests/unit/services/test_bootstrap_config_refs.py
@@ -2,8 +2,10 @@ from pathlib import Path
 
 import pytest
 
+from datapipeline.build.config_hash import compute_config_hash
 from datapipeline.services.bootstrap.config import _globals, _load_by_key, artifacts_root
 from datapipeline.services.bootstrap.core import _load_sources_from_dir
+from datapipeline.services.project_paths import sources_dir
 
 
 def _write_project_yaml(
@@ -139,3 +141,67 @@ def test_missing_env_ref_raises_clear_error(tmp_path: Path, monkeypatch: pytest.
 
     with pytest.raises(ValueError, match="MISSING_ROOT"):
         _globals(project_yaml)
+
+
+def test_sources_dir_resolves_env_backed_project_path_aliases(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    monkeypatch.delenv("RAW_ROOT", raising=False)
+    project_root = tmp_path / "project"
+    raw_root = tmp_path / "shared"
+    (raw_root / "sources").mkdir(parents=True)
+    _write_project_files(project_root)
+    (project_root / ".env").write_text(f"RAW_ROOT={raw_root}\n", encoding="utf-8")
+    project_yaml = _write_project_yaml(
+        project_root,
+        globals_lines=["raw_root: ${env:RAW_ROOT}"],
+    )
+    project_yaml.write_text(
+        project_yaml.read_text(encoding="utf-8").replace(
+            "  sources: sources",
+            "  sources: ${raw_root}/sources",
+        ),
+        encoding="utf-8",
+    )
+
+    assert sources_dir(project_yaml) == (raw_root / "sources").resolve()
+
+
+def test_compute_config_hash_changes_when_env_value_changes(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    monkeypatch.delenv("RAW_ROOT", raising=False)
+    project_root = tmp_path / "project"
+    _write_project_files(project_root)
+    (project_root / "sources" / "sample.yaml").write_text(
+        "\n".join(
+            [
+                "id: sample.fs",
+                "parser:",
+                "  entrypoint: identity",
+                "  args: {}",
+                "loader:",
+                "  entrypoint: core.io",
+                "  args:",
+                "    transport: fs",
+                "    format: jsonl",
+                "    path: ${raw_root}/rows.jsonl",
+            ]
+        )
+        + "\n",
+        encoding="utf-8",
+    )
+    project_yaml = _write_project_yaml(
+        project_root,
+        globals_lines=["raw_root: ${env:RAW_ROOT}"],
+    )
+
+    monkeypatch.setenv("RAW_ROOT", "/tmp/raw-one")
+    hash_one = compute_config_hash(project_yaml, project_root / "tasks")
+
+    monkeypatch.setenv("RAW_ROOT", "/tmp/raw-two")
+    hash_two = compute_config_hash(project_yaml, project_root / "tasks")
+
+    assert hash_one != hash_two
